More Settings Options, Smoother Animations, and Bug Fixes

+++ Assets/Scripts/Interfaces/ISound.cs
@@ -1,4 +1,4 @@
 ﻿interface ISound
 {
-    void UpdateMaxVolume(float newVolume);
+    void UpdateMaxVolume(int newVolume);
 }
